% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboardAnalysis-package.R
\docType{package}
\name{billboardAnalysis-package}
\alias{billboardAnalysis}
\alias{billboardAnalysis-package}
\title{billboardAnalysis: Song Billboard Analysis in the 2010s}
\description{
The R packages integrates Billboard chart data and Spotify audio features into an interactive Shiny dashboard. This dashboard allows users to explore song trends from the 2010s, focusing on audio characteristics like energy, acousticness, danceability, and speechiness. It features interactive plots, including year-wise trends of audio features, feature distributions for long-lasting tracks, and comparisons of features across genres. The dashboard is designed to provide insights into the evolution of music features over time, with a user-friendly interface and help mode for navigating the features.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://etc5523-2024.github.io/assignment-4-packages-and-shiny-apps-bigchayanont/}
}

}
\author{
\strong{Maintainer}: Chayanont Piyawatcharvijit \email{big.chayanont@gmail.com}

}
\keyword{internal}
