% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\title{Launch Billboard Analysis Shiny App}
\usage{
run_app()
}
\value{
Nothing
}
\description{
This function launches a Shiny application designed to provide interactive visualizations and analysis of Billboard and Spotify data. Users can explore audio features, chart trends, and other insights related to popular music.
}
\examples{
\dontrun{
run_app()
}
}
